refactor: move fragment to each package
