+++ app/src/main/java/com/itinergo/ui/account/DetailProfileFragment.kt
@@ -1,0 +1,22 @@
+package com.itinergo.ui.account
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.itinergo.R
+
+
+class DetailProfileFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_detail_profile, container, false)
+    }
+
+
+}
