+++ app/src/main/java/com/itinergo/ui/findtrip/AddTripFragment.kt
@@ -1,4 +1,4 @@
-package com.itinergo.ui
+package com.itinergo.ui.findtrip
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,15 +7,14 @@ import android.view.View
 import android.view.ViewGroup
 import com.itinergo.R
 
-
-class TravelTipsFragment : Fragment() {
+class AddTripFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_travel_tips, container, false)
+        return inflater.inflate(R.layout.fragment_add_trip, container, false)
     }
 
 }
