+++ app/src/main/java/com/itinergo/ui/findtrip/VoilaFtripFragment.kt
@@ -1,0 +1,20 @@
+package com.itinergo.ui.findtrip
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.itinergo.R
+
+class VoilaFtripFragment : Fragment() {
+
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_voila_ftrip, container, false)
+    }
+}
