+++ app/src/main/java/com/itinergo/ui/home/PlaceTodayFragment.kt
@@ -1,4 +1,4 @@
-package com.itinergo.ui
+package com.itinergo.ui.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
