+++ app/src/main/java/com/itinergo/ui/place/PlaceFragment.kt
@@ -1,0 +1,20 @@
+package com.itinergo.ui.place
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.itinergo.R
+
+class PlaceFragment : Fragment() {
+
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_place, container, false)
+    }
+}
