+++ app/src/main/java/com/itinergo/ui/place/VoilaPlaceFragment.kt
@@ -1,4 +1,4 @@
-package com.itinergo.ui
+package com.itinergo.ui.place
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,16 +7,13 @@ import android.view.View
 import android.view.ViewGroup
 import com.itinergo.R
 
-
-class DetailProfileFragment : Fragment() {
+class VoilaPlaceFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_detail_profile, container, false)
+        return inflater.inflate(R.layout.fragment_voila_place, container, false)
     }
-
-
 }
