+++ app/src/main/java/com/itinergo/ui/savedplace/DetailSavedPlaceFragment.kt
@@ -1,0 +1,21 @@
+package com.itinergo.ui.savedplace
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.itinergo.R
+
+
+class DetailSavedPlaceFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_detail_saved_place, container, false)
+    }
+
+}
