+++ app/src/main/res/layout/fragment_detail_profile.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:ignore="HardcodedText,ContentDescription"
-    tools:context=".ui.DetailProfileFragment">
+    tools:context=".ui.account.DetailProfileFragment">
 
 
     <ImageView
