+++ app/src/main/res/layout/fragment_detail_travel_tips.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.DetailTravelTipsFragment">
+    tools:context=".ui.traveltips.DetailTravelTipsFragment">
 
 
     <ImageView
