+++ app/src/main/res/layout/fragment_place.xml
@@ -5,7 +5,7 @@
     android:id="@+id/tv"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.PlaceFragment">
+    tools:context=".ui.place.PlaceFragment">
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/textInputLayout"
