+++ app/src/main/res/layout/fragment_place_today.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.PlaceTodayFragment">
+    tools:context=".ui.home.PlaceTodayFragment">
 
     <ImageView
         android:id="@+id/img_place_today_detail"
