+++ app/src/main/res/layout/fragment_saved_place.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.SavedPlaceFragment">
+    tools:context=".ui.savedplace.SavedPlaceFragment">
 
 
     <androidx.recyclerview.widget.RecyclerView
