+++ app/src/main/res/layout/fragment_travel_tips.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.TravelTipsFragment">
+    tools:context=".ui.traveltips.TravelTipsFragment">
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/textInputLayout"
