+++ app/src/main/res/layout/fragment_voila_budget.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background"
-    tools:context=".ui.VoilaHomeFragment"
+    tools:context=".ui.home.VoilaHomeFragment"
     tools:ignore="HardcodedText, ContentDescription">
 
     <ImageView
