+++ app/src/main/res/layout/fragment_voila_home.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/background"
     tools:ignore="HardcodedText, ContentDescription"
-    tools:context=".ui.VoilaHomeFragment">
+    tools:context=".ui.home.VoilaHomeFragment">
 
     <ImageView
         android:id="@+id/splashscreen_bg2"
