+++ app/src/main/res/layout/fragment_voila_saved_place.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/purple"
-    tools:context=".ui.VoilaPlaceFragment">
+    tools:context=".ui.place.VoilaPlaceFragment">
 
 
     <ImageView
