+++ app/src/main/res/layout/place_list_item.xml
@@ -9,7 +9,6 @@
         android:layout_width="282dp"
         android:layout_height="81dp"
         android:background="@drawable/bg_place_list"
-        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.24"
         app:layout_constraintStart_toStartOf="parent"
